Move inspection module creation to the workspace area
